+++ src/include/stack.h
@@ -8,8 +8,8 @@ typedef struct STACK_STRUCT
     unsigned char* data;
     size_t size;
     size_t init_size;
-    u_int64_t ptr;
-    u_int64_t base_ptr;
+    u_int32_t ptr;
+    u_int32_t base_ptr;
 } stack_T;
 
 stack_T* init_stack(size_t init_size);
@@ -17,9 +17,9 @@ stack_T* init_stack(size_t init_size);
 void stack_push(stack_T* stack, char data);
 unsigned char stack_pop(stack_T* stack);
 
-void stack_set_value(stack_T* stack, u_int64_t ptr, unsigned char data);
-unsigned char stack_get_value(stack_T* stack, u_int64_t ptr);
+void stack_set_value(stack_T* stack, u_int32_t ptr, unsigned char data);
+unsigned char stack_get_value(stack_T* stack, u_int32_t ptr);
 
-void stack_set_frame(stack_T* stack, u_int64_t ptr);
+void stack_set_frame(stack_T* stack, u_int32_t ptr);
 
 #endif
